requirements.txt
streamlit
transparent-background
pillow
tqdm